Add a separate pre-commit config specifically for GitHub Actions
This config has the terraform_validate hook commented out, since
terraform validate cannot currently run without accessing the remote
state because of the way the providers are defined in
terraform/providers.tf.  This is something that may be remedied in the
future.  For more information, check out these two GitHub issues:
* https://github.com/hashicorp/terraform/issues/15895
* https://github.com/hashicorp/terraform/issues/15811