Add debugging output to the loop the initializes Terraform directories.
I was looking at the output of a failing GitHub Actions job and realized that
the loop doesn't give an indication of what directory it's trying to initialize.
I think it's a good idea for debugging to know what has succeeded and where it
is failing for troubleshooting purposes, so I'm adding an `echo` to indicate
where it's trying to work.