Disassociate the TGW VPC attachment and the TGW's default route table
We do this using the AWS CLI via a null_resource with a local-exec
provisioner.  It would be nice if we didn't have to use this escape
hatch, but Terraform does not support this sort of cross-account foo
right now.

Note that after asking AWS to disassociate the transit gateway VPC
attachment and the transit gateway's default route table we sleep in a
loop until the disassociation is complete.