Make all instances that mount the EFS volume depend on the EFS mount target(s)
We have been seeing cases, particularly among the Pentest Portal
instances, where the EFS volume does not get mounted at boot.
Afterwards, a simple mount -a command does successfully mount the EFS
volume.

In my attempts to reproduce the problem, I found that it was possible
for the EFS mount target(s) to be created _after_ the EC2 instances
that require it to mount the EFS volume.

To remedy this I added a depends_on = [aws_efs_mount.target.target] to
such instances, and it appears to get rid of the problem; at least, I
am unable to reproduce the problem after making this change.

Hence I am willing to state that this change resolves #85.