Add a cloud-init shell script to ensure mounting of the EFS share
Even with the changes in commit
638f1824d588b0631048ba3437e72d882da1349c, it is still possible for an
instance so slip through the cracks and boot without mounting the EFS
share.  This is because Terraform has no way to distinguish between
the "resource created" and "resource created and ready for use"
states.  The share can be created, but not become available for
attachment until _after_ the instance has already booted up and
automounted the mounts listed in /etc/fstab.