Specify the region name and region-specific endpoint URL for AWS CLI STS calls
STS used to be un-regioned, like S3, but now it is regioned.  The AWS
CLI by default uses the global endpoint URL https://sts.amazonaws.com
instead of the region-specific one, even when you specify the region.
Therefore, we have to set the region-specific endpint URL manually.

This is important since the STS VPC endpoint _only_ sets a local DNS
record to override the _local region's_ public STS endpoint.  If we
don't specify the endpoint URL then the AWS CLI will reach out to the
_global_ https://sts.amazonaws.com URL, and that DNS entry will still
point to an external IP.