Rework inbound ports locals into a single variable
This single variable combines all the ports specified in the inbound_ports_allowed variable, removes duplicates, and create a dictionary suitable for use in a for_each Terraform statement.

Note that we set the key of the resulting map to the index of the port in the flattened and deduped list.  This index is convenient when numbering the corresponding ACL rules for the operations subnet.