Bump up the size of the Debian Desktop instances
@dav3r noticed that a Debian Desktop instance had locked up this
morning.  Upon inspection of the kernel logs I saw that it had run out
of memory.  I noticed that these instance types were only t3.small,
which has a mere 2GB of RAM.  That's a bit thin on memory for running
Linux with a desktop environment and a modern web browser, so I bumped
the instance type up to a t3.medium.  These instance types have 4GB of
RAM, which should suffice.

Co-authored-by: dav3r <david.redmin@trio.dhs.gov>