Add an input to the cert installation script
This Boolean input determines whether or not the destination
directories for the full chain and private key PEMs are created.  In
the case of the Guacamole instance it makes sense to do this, but
normally you would want the script to error out if the destination
directory does not exist.