Put Guacamole in the same subnet as Guacamole for now
For some reason I can't SSM to the instance unless I put it in the
first private subnet.  I believe this has something to do with the
NACLs that are in place for that subnet specifically.  I will figure
this out later.  I'd definitely prefer to put this instance in a
different subnet than the Guacamole instance.