Add a rule to allow the private subnets to ingress via port 80
This is necessary so that they can egress to the internet via port 80
to install packages.

Also fix the comments for the HTTPS ingress and ephemeral port egress
ACL rules.  These comments used to describe how the rules were
necessary for the Guacamole instance to download its Docker image, but
they are now only necessary for the Terraform instance to install
packages and perform terraform init.