Add a rule allowing egress anywhere via HTTP
This is necessary for the Terraformer instance to update and install
packages.

Also fix a comment on the HTTPS egress rule that mentions it being
necessary for the Guacamole instance to download its Docker image.
This is no longer the case, but it _is_ necessary for the Terraformer
instances to perform terraform init.