Remove count where unnecessary on Nessus resources
I noticed that we were creating identical policies and roles for each
Nessus instance, but then using only the zeroth one, so this change
gets rid of some duplicate resources.  It also makes it much easier
for me to make the Nessus instance role an output so that it can be
used by the Terraformer instance.