Read in the connection SQL template once
Adjust the script that renders SQL templates for Guacamole connections to only
read in the SQL template once instead of for every connection. This data is
used as-is with no changes made to it so it doesn't make sense to read in the
template data for every connection. Instead we read it in once and use the same
template data for each connection's generated file.