Use SSM Parameter Store parameters
Replace the statically declared values for the owner/group used in the
efs-mount cloud-config template file with values from SSM Parameter Store
parameters. Since these values are used when building the AMIs used in this
Terraform configuration it will allow this configuration to use the correct
values when it is applied.

Co-authored-by: dav3r <david.redmin@trio.dhs.gov>