Update the EFS access point configuration
Add the root_directory argument to the EFS access point configuration
with settings to ensure that if no root directory exists, then the
created root directory will match the UID/GID settings of the
posix_user argument. We need to specify a non-root path in the EFS to
serve as the root directory for the access point so we can ensure that
permissions are corrected when used to mount the share locally. If we
do not specify a non-root path the root of the EFS will be used, and it
has root:root permissions by default.