Inject a dependency of the session manager foo on the corresponding policy
This is done by copying a variable together with its default value
over from cisagov/session-manager-tf-module so that it its value can
be specified outside of that module.  This allows us to impose a
dependency of the module on the policy that allows for the creation of
its resources; otherwise, this dependency will be cyclical.