Bump up size of pentestportal instances
These PTP instance types were t3.small, which has a mere 2GB of RAM.  That's a bit thin on memory for running Linux with a desktop environment and a modern web browser, so I bumped the instance type up to a t3.medium.  In the past, we have seen fairly regular occurrences of these PTP instances locking up, which is likely due to lack of available memory.  These t3.medium instances have 4GB of RAM, which should improve our situation.

A similar change was made in #106 for our Debian Desktop instances.