Validate assessment ID and type variables
This is a semi-hacky way of validating user-provided assessment information.  As of Terraform v1.3.x, there is no built-in way to validate one variable against another, as we need to do here.  Terraform input variable
validation expressions can only refer to the variable being validated and no others.  For more, see https://developer.hashicorp.com/terraform/language/expressions/custom-conditions#input-variable-validation

To work around this limitation, we use a null_resource with a lifecycle block that contains a precondition.  The precondition is evaluated at plan time, and if it fails, the plan will fail with the error message provided.