Rename Assessor Portal to Assessor Workbench
It is necessary to do this now, since we recently updated Guacamole to
use ED25519 SSH keys.  The old Assessor Portal AMI is no longer being
created, and hence it cannot support the new SSH key type.  Therefore
it is imperative to start using the new Assessor Workbench AMI now.