Overwrite authorized_keys on Windows instances
The public SSH key currently built into the Windows AMIs is no longer
used. Since we are in the process of migrating Windows AMI creation we
are trying to avoid the manual conversion process that was used to
create the AMIs currently in use. This patch to EC2Launch will provide
the correct SSH public key so that Guacamole can connect to Windows
instances over sftp.