Add a cloud-init fragment to reapply the netplan config
The Netplan config is created by cloud-init, but it needs to be
slightly modified and reapplied.  There is a Python script that fixes
the DHCP4 options in the netplan configuration already generated by
cloud-init.  This cloud-init fragment simply reapplies the Netplan
configuration after the modification.

See these issues for more details:
- cisagov/skeleton-packer#300
- canonical/cloud-init#4764