Temporarily use a non-default branch of cisagov/terraform-state-read-role-tf-module
This is being done for testing purposes.  We can revert to the default
branch of cisagov/terraform-state-read-role-tf-module once
cisagov/terraform-state-read-role-tf-module#40 is merged.

The changes in cisagov/terraform-state-read-role-tf-module#40 allow us
to specify a list of additional remote states to which the Terraform
read/write role should also be granted access.  This is necessary to
grant sufficient permissions to those assessors who will be allowed to
reprovision their own assessments.