Remove needless quotes from YAML string that is a regex
Also change the string to be a multiline YAML string.

Both these changes are necessary to avoid errors from
yamllint/ansible-lint.  Note that changing the string to be a
multiline YAML string changes the rules around escaping special
characters.

I deployed a gophish instance to our dev-a COOL environment to verify
that the modified regex continues to function as expected.